templates/graphql_template/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
slowapi==0.1.9
cryptography==45.0.6
requests==2.31.0
pydantic==2.11.4
strawberry-graphql[fastapi]==0.235.2
graphql-core==3.2.3
typing-extensions==4.12.2